requirements.txt
accelerate==1.9.0
captum==0.8.0
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
filelock==3.13.1
fonttools==4.59.0
frozendict==2.4.6
fsspec==2024.6.1
graphviz==0.21
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.0.0
pipreqs==0.4.13
psutil==7.0.0
pyparsing==3.2.3
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
scipy==1.16.0
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchdiffeq==0.2.5
torchvision==0.21.0+cu126
torchviz==0.0.3
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.5.0
yarg==0.1.10
